# Dealer Adapt API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Dealer Adapt API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Favorite/{id}": {
      "get": {
        "tags": [
          "Favorite"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Favorite"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Favorite": {
      "get": {
        "tags": [
          "Favorite"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Favorite"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Favorite"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Password/forgot": {
      "post": {
        "tags": [
          "Password"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ForgotPasswordModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Password/reset": {
      "post": {
        "tags": [
          "Password"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ResetPasswordModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Favorite": {
      "type": "object",
      "properties": {
        "id": {
          "format": "uuid",
          "type": "string"
        },
        "carId": {
          "format": "uuid",
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "dateTime": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "ForgotPasswordModel": {
      "type": "object",
      "properties": {
        "email": {
          "format": "email",
          "type": "string"
        }
      }
    },
...
    "ResetPasswordModel": {
      "type": "object",
      "properties": {
        "email": {
          "format": "email",
          "type": "string"
        },
        "token": {
          "type": "string"
        },
        "newPassword": {
          "format": "password",
          "type": "string"
        }
      }
    }
  }
}